[deserialization] When deserializing vtables, do not deserialize the functions they reference.
This fixes a bug where we were deserializing a function with a call to a shared
linkage function. The shared linkage function was never deserialized causing an
assertion to fire due to shared linkage functions always needing a definition.

I am planning on implementing lazy deserialization of vtable functions.

Swift SVN r14581