Don't allow a subobject initializer to delegate to another initializer.
While it would be technically sound to do this, it almost surely means
that the user forgot the '-> Self' to make this a complete object
initializer. Provide a Fix-It to add the '-> Self'.

If we ever feel like we need this behavior, we can introduce a special
syntax for it.


Swift SVN r14598