[serialization] Be more careful about deserializing the same Decl twice.
If we got just the wrong chain of dependencies, we'd create two VarDecls
for the same variable. I'm not sure if any of the other decls have the
same problem, but better safe than sorry.

I couldn't think of a small test case for this, but it does fix the one
in the Radar.

<rdar://problem/16205936>

Swift SVN r14625