Coerce the self argument of @objc methods to the appropriate type.
When lowering @objc entrypoints, we need to set the argument explosion
to a value with the proper type for use within the body of the function.

Change r13931 worked around the lack of proper entrypoint lowering by
adding a cast at a use site. Now that this is fixed we can remove that
work-around.

Fixes <rdar://problem/16075395>.

Swift SVN r14666