implement: <rdar://problem/16193162> Require specifying self for locations in code where strong reference cycles are likely
This requires that property references have an explicit "self." qualifier when in an
explicit closure expression, since self will be captured, not the property.

We don't do the same for autoclosures.  The logic here is that autoclosures can't 
practically be used in capturing situations anyway, since that would be extremely 
surprising to clients.  Further, forcing a syntactic requirement in an autoclosure
context would defeat the whole point of autoclosures: make them implicit.


Swift SVN r14676