When running with -sil-verify-all only verify the function processed by a function pass after every iteration instead of the whole module.
Verifying the module after every function transform is run on every function
takes so much compile time that the mode is not really useful. I don't imagine
we are really getting anything for verifying the whole module so it makes sense
to limit what we are verifying.

Swift SVN r14834