SILParser: use a Scope for generic parameters.
This will fix the scoping issue for generic paremeters in SILType.
We use a Generics scope when parsing generic paremters of a SILType, after
we are done parsing the type, we delete the Generics scope.

Testing case will be added when SILParser can correctly handle generic code.

rdar://14443287


Swift SVN r14998