fix rdar://16280138 - synthesized getter is defined in terms of archetypes, not interface types
We have some unfortunate representational problems in the AST for function 
return types that was biting getters synthesized for properties.  Work around this by
delicately doing a mapping late in the one place where it matters in the generics sytem.

Many thanks to Doug for writing this patch. :-)



Swift SVN r15003