SILParser: parsing generic code.
The '.Archetype' property of Substitution is not used and SILPrinter
prints a different name for Substitution and for the generic param list.

The fix is to ignore the '.Archetype' property in SILParser and instead
to match up Substitution with the archetype by ordering alone.

The silgen output from generic_closures.swift can now be parsed with SILParser
with changes:
  typealias Int = Int64 to typealias Int = Builtin.Int64
  typealias Char = Int32 to typealias Char = Builtin.Int32
Without the changes, SILParser will complain about mismatching Int64
with Builtin.Int64.

The next step is for SILPrinter to not print '.Archetype' property of
Substitution. The majority of work is updating testing cases.

rdar://14443287


Swift SVN r15014