Add prims to Benchmark.swift.
We are actually doing a pretty good job optimizing this (even though I
am sure it could be better). The test case is a 5000 node prim run. IIRC
when I ran the cpp version of this on my local cpu it was instantaneous.
The swift version is ~1.5 seconds.

Swift SVN r15164