Correctly index into the parameters for partial_apply.
We need to adjust the index by the number of actual arguments used in
the partial apply.

There's no way to test this at this time as the only pass using this is
allocbox-to-stack, and that pass ends up bailing out due to the
ObjectPointer for the box also getting passed to the partial_apply.

No diffs compiling stdlib.

Swift SVN r15183