Change the prims cpp benchmark to match the prims swift benchmark.
Also up the number of iterations the swift benchmark in the stdlib runs so that
it matches the number of iterations the c++ version does. This iteration number
is enough to cause the c++ version to run for ~1 second. With this same number
of iterations, the swift version takes ~10 seconds.

Swift SVN r15205