Add Type::visit to formalize visiting a type and all of its children instead of having code outside of Type abuse findIf/transform.
There are many places in the code base where people are abusing
Type::findIf/Type::transform just to visit a type and all of its
subtypes. This is an abuse of those methods and is endlessly confusing
especially when someone knew to this area of the code base reads
it.

Type::visit just encapsulates that behavior in a common sense name so
that the writers can be clear about their purpose and readers can be
sure what the writer is attempting to do without questions.

Swift SVN r15276