TypeChecker: Remove the notion of "external decls we've already handled".
There's no reason to check these before everything else. We'll still check
them in the main processing loop of performTypeChecking.

(There are cases where we create a type checker outside of
performTypeChecking, but the existing code doesn't handle that properly
either -- it doesn't call handleExternalDecl for new decls being brought
in. If this turns out to be an issue, we can restore the loop and fix that
issue.)

Swift SVN r15353