Start wiring up @availability checking for types.
As a use case, declare a new 'NSUInteger' (which the importer maps to 'Int'
when importing from Objective-C) and mark it unavailable with a
helpful message.  This will eventually go to the Foundation overlay
once @availability is properly serialized.

There's some duplication in availability checking.  That will get
consolidated as more of it gets written.  It's not very comprehensive
right now.

Swift SVN r15583