Always coerce the operand of an optional force to Optional type,
The attempt to coerce the operand to UncheckedOptional for
dynamic member accesses leads to an inconsistency when
applying the constraint solution because we haven't figured
out how to convert to UncheckedOptional<T> --- all the
conversion restrictions are to optional type.

The test case will be part of the standard test suite when
applying the UncheckedOptional import patch.

In order to continue to work around weaknesses in SILGen's
function conversion code, peephole optional conversion in
the function application path.

Swift SVN r15651