Collect the box pointer operands of partial_apply that are unneeded.
For the partial_apply instructions we'll rewrite to use stack operands,
we need to remove the box pointer operands when specializing the applied
functions and rewriting the partial_apply.

This is another commit with no current effect - we do not specialize
yet, and do not promote any additional boxes to stack locations.

Swift SVN r15719